+++ gql/lib/src/validation/validating_visitor.dart
@@ -1,4 +1,6 @@
 import "package:gql/ast.dart";
 import "package:gql/src/validation/validator.dart";
 
-abstract class ValidatingVisitor extends SimpleVisitor<List<ValidationError>> {}
+abstract class ValidatingVisitor extends SimpleVisitor<List<ValidationError>> {
+  const ValidatingVisitor();
+}